updating auth on user and grocery list routes

+++ src/controllers/user.controller.js
@@ -1,38 +1,47 @@
 import models from '../models/index.js';
 import { genericUpdateSuccessResponse, signToken, serializeList, isAdminOrOwner } from '../utils/utils.js';
+import { isAdmin } from '../utils/utils';
 
 const { User } = models;
 
 const getAllUsers = async (req, res, next) => {
 	const offset = req.query.offset || 0;
 	const limit = req.query.limit || 25;
-	try {
-		const users = await User.findAll({ offset, limit });
-		res.json(serializeList(users));
-	} catch (err) {
-		next(err);
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
+			const users = await User.findAll({ offset, limit });
+			res.json(serializeList(users));
+		} catch (err) {
+			next(err);
+		}
 	}
 };
 
 const getUserById = async (req, res, next) => {
-	try {
-		let user = await User.findOne({ where: { uuid: req.params.userId } });
-		res.json(user.serialize());
-	} catch (err) {
-		next(err);
+	if (!isAdminOrOwner(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
+			let user = await User.findOne({ where: { uuid: req.params.userId } });
+			res.json(user.serialize());
+		} catch (err) {
+			next(err);
+		}
 	}
 };
 
 const updateUser = async (req, res, next) => {
-	try {
-		if (isAdminOrOwner(req)) {
+	if (!isAdminOrOwner(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
 			let response = await User.update(req.body, { where: { uuid: req.params.userId }, returning: true });
 			genericUpdateSuccessResponse(response, res);
-		} else {
-			res.sendStatus(403);
+		} catch (err) {
+			next(err);
 		}
-	} catch (err) {
-		next(err);
 	}
 };
 
@@ -59,11 +68,15 @@ const login = async (req, res, next) => {
 };
 
 const deleteUser = async (req, res, next) => {
-	try {
-		await User.destroy({ where: { uuid: req.params.userId } });
-		res.sendStatus(200);
-	} catch (err) {
-		next(err);
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
+			await User.destroy({ where: { uuid: req.params.userId } });
+			res.sendStatus(200);
+		} catch (err) {
+			next(err);
+		}
 	}
 };
